+++ src/test/java/org/morphix/lang/ComparablesTest.java
@@ -1,0 +1,23 @@
+package org.morphix.lang;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.equalTo;
+
+import org.junit.jupiter.api.Test;
+import org.morphix.reflection.Constructors;
+import org.morphix.utils.Tests;
+
+/**
+ * Test class for {@link Comparables}.
+ *
+ * @author Radu Sebastian LAZIN
+ */
+class ComparablesTest {
+
+	@Test
+	void shouldThrowExceptionOnCallingConstructor() {
+		UnsupportedOperationException unsupportedOperationException = Tests.verifyDefaultConstructorThrows(Comparables.class);
+
+		assertThat(unsupportedOperationException.getMessage(), equalTo(Constructors.MESSAGE_THIS_CLASS_SHOULD_NOT_BE_INSTANTIATED));
+	}
+}